[refactor] SearchExhList 컴포넌트에서 ExhItem 분리
